set initial tests and environment
Here is the initial tests and environment. We will use Mocha.js
and should.js for test framework and assertion library respectively.

We might need to read the following links to understand tests in
node.js and nvm.

https://quickleft.com/blog/creating-and-publishing-a-node-js-module/
http://mochajs.org/
https://github.com/shouldjs/should.js

Also, for the testing, run

$ make test

or

$ nvm test

This is the initial tests so tests are passed intentionally without
meaningful things. So, you need to revise each test cases.