fixed: data[i].assignee makes db pending because it is not string type.
cherry-picked from issue9 branch

Based on schema, assignee should be type of text, but data[i].assignee
is not string but json object. Interestinglly, many of issues from test
repository(github/android) does not have assignee, so it's OK because
it means the value is null.

However, there has some issues having assignee which is not string but
object. So, in this case, Cassandra raises error and makes clients are
pending. (Reject clients and they are pending)

So, I just make a workaround for it.

fixed #9