no more than limit git clone jobs will be simultaneously running
Too many git clone simultaneously can make perfomance of system
significantly bad, because that's not only I/O but also CPU bound
job. So, settting limit by number of cores * 2 is good for other
tasks on the system.

However, jobs of async.eachLimit stop when error happens. So, need
to be careful about it.

https://github.com/caolan/async#eachlimitarr-limit-iterator-callback