Do searchRepos for github on everyday at 00:00:01am
This does crawling information of github repos on everyday morning
by node-cron. Everyting is good, but one downside, so far, is that
it starts another job even the prior one has not finised yet.

So, now, I expect there has no problem, because crawling information
of repos from github takes at most 6 hours for 5 years, which is
from 2011-01-01 to today (2015-06-19).

However, we need to be aware of it.

Reference:
https://github.com/ncb000gt/node-cron